+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased]
+
+## [0.1.0] - 2025-07-21
+### Added
+- Initial dataset builder and CLI
+- Unit tests and CI configuration
+
 ## [3.0.0] - 2025-07-15
 
 ### Added